build: do not inline resources for normal builds
* Inlining the resources for normal builds does cause issues with the livereload.

For example if I change a HTML file, it will be properly copied over (per assets gulp task), but it's still inlined wrong and does not update the view.

By default all build tasks should not inline the resources, and only the build tasks (like with the ng compiler) should inline the resources.