fix(build): binary execution on windows
Currenlty the `execNodeTask` helper function resolves the binary from the `package.json` and spawns a new process of it.

Specifically on Windows this will break, because the Windows Shell does not know how to run the given files (from the Shebang or PATHEXT)

Since those binaries are always node binaries we could manually run the node interpreter and run the given JS file as an argument.

Closes #1129