feature(md-input): refactor MdInput as an attribute
The goal is to simplify the MdInput interface and allow it to evolve and
collaborate with other directives on an <input>. For example, Chips
could be a directive on the `<input>` as well.

The next steps are going to be adding back MdPlaceholder as such a
directive that collaborates with MdInput.