fix(input): prevent overlapping placeholder on inputs with native masking
Prevents the Material placeholder overlapping with native input masking (e.g. on `input[type="time"]`.
Previously, this was hardcoded for `date`, however it didn't take into account that not all browsers support `date` inputs.
This change runs a feature test on each of the types that should have input masking and only triggers the special behavior if the type is supported.

Fixes #1853.