fix character count hint for input.
fix (input): Remove the property CharacterCounter on md-input. Add name
of local variable on md-input and add md-hint to fix counter binding of
md-input for the show hint lable with charater count example.

The property characterCounter does not exist on md-input and therefore
the character count was not being displayed for the show hint lable with
charater count example. Removing this and adding the name
#characterCountHintExampleInput along with the md-hint tag using the
name and the character count property fixes the issue.