perf(tabs): reduce amount of reflows when aligning the ink bar
Currently the ink bar gets aligned to the active tab every time change detection runs. These changes switch to aligning it if the selected index has changed or the header's content has changed. This can be improved further if we were to debounce the MutationObserver changes (will add in a follow-up PR).