fix(input): work around Chrome issue that collapses the input
This fix works around a Chrome bug, where number inputs that have a `min` and a `max`, but not a `step`, will collapse to about 50% their proper width. This probably isn't a huge use case, but it's probably good to have the inputs be consistent across browsers.

For more info the the bug: https://bugs.chromium.org/p/chromium/issues/detail?id=598164

Fixes #2470.