fix(slide-toggle): input not updated after drag
The `MdSlideToggle` uses the OnPush ChangeDetection strategy for the component.

**Problem Explanation**

When the thumb is being dragged and the value changes, the new value is currently applied after a timeout.

This timeout does not trigger any change detection, and the new value won't be applied to the underlying input.

**Solution**

With this change, the new value is applied immediately in the `dragend` event (because events trigger a change detection) and just the `dragging` state will be updated in the next tick.

The timeout is required to ensure that there will be no `click` event after the `dragend` event.