fix(select): missing rxjs filter import
The `filter` RxJs operator is being used inside of the select.  Building the select package on its own would not work because the `filter` operator is not present.

The failures just don't show up because in some other packages we import the `filter` operator and it will be applied to the global Observable prototype.