build: silence build and test script
Currently the CI traces everything that runs inside of the bash script.

This might have been useful when creating the script but now it really pollutes the CI output which should be clean and structured.

Disabling tracing for the build and test script because the errors will still show up and the CI will look more clean.