Alphabetical sorting for Feature Status List
The feature status list in the project's README.md seems to be sorted by feature status (or something else) right now which is not very helpful when trying to find the status of a component one cares about. Alphabetical sorting would make it easier to identify the feature you are looking for.