build: get previous payload by querying git
Currently the previous payload is just the latest uploaded payload result on Firebase. This is problematic because it can happen that the current payload is uploaded "too fast" and the payload task is comparing against itself.

Just calculating before uploading could work most of the time, but there it would still fail if another Travis Job uploades payload results at the same time.

The safest solution would be resolving the parent commit of the current Travis job. This way the payload diff would be always correct (even if multiple Jobs are running).