build: payload task fails if functions are down
It can happen that the firebase `payloadGithubStatus` function is temporary disabled because a deploy failed (for example due to deploying while function executes).
This leads to a failure inside of the payload CI task because the task is trying to parse the body using JSON (which is not always a JSON response)

It's not worth trying to parse the JSON value since it's only used for debugging and to give an insight what's going on in the functions.