fix(input): underline showing at end if text-align is set
Recently the `align` input binding has been removed in favor of the CSS property `text-align`.

If a developer sets the `text-align` property to `end` the text will start from the end and also the input underline will show incorrectly start from the end.

Setting the absolute positioned underline to `left: 0` ensures that the underline always shows-up correctly (because the `mat-input-underline` is inside of a relative container)

Fixes #5272