chore: remove ignored docs-private tags
With #5314 a lot of `@docs-private` tags that are ignored have been added. Dgeni only includes exports of files if they are explicitly exported in a `src/XXX/index.ts` file.

Since this is not the case the `@docs-private` tags can be removed because the Rx chain files are not re-exported anywhere.