chore: build cdk and material in npm build task
Apparently developers that clone the Material repository to get their own version of Material are running the `npm run build` command and are expecting the CDK and Material release output to be inside of the `dist/` folder.

Currently only the Material package will be built and the CDK isn't built at all. This commit changes the npm build task so that it builds every release package of the publish task.

Closes #5372