build: livereload not working for cdk changes
When serving the demo-app the Material package will be watched using the generated `material:watch` task. This does not mean that the `cdk:watch` task will run too because Material depends on the CDK.

With this change, the watch task will now also run the watch tasks of the packages it depends on.