fix(connected-position-strategy): multiple places to specify direction
* Currently, if the consumer wants to handle layout direction in a connected overlay, they have to specify it in two places: via the `OverlayState.direction` and the `ConnectedPositionStrategy.withDirection`. This can be confusing and easy to overlook. These changes remove the `ConnectedPositionStrategy.withDirection` method, in favor of passing in the overlay state direction.
* Makes a few underscored properties private since they weren't being used anywhere outside of the `ConnectedPositionStrategy` class.

BREAKING CHANGE: This will be a breaking change if somebody was using the `withDirection` method and the way to fix it is to only pass the direction to the `OverlayState`. If we want to be safer about it, I can put the method back in and mark it as deprecated.