refactor: use cdk imports instead of re-exports
Recently a bunch of exports have been moved from Material to the CDK and to avoid a lot of changes the imports for those exports stayed the same.

This is possible because every moved export was just re-exported in the original source file again.

Now the imports are switched to the real `@angular/cdk` imports and the unnecessary files inside of `core/` have been removed (for clarity)