build: artifact scripts should push empty commits
When publishing build artifacts (cdk; material) there will be always a change (regardless if something changed in the package or not). This is because the version will be always appended with the current SHA.

But when publishing the docs content this is not the case. Sometimes there won't be any change and Git will exit with error code `1` because empty commits are not allowed.

In the future we might also just append the SHA to the version of the `material-examples` package.