build: faster live-reload for theme changes
Whenever something inside of a theme changes and the demo-app is served, the watch task waits for the CSS files to be built. This is superfluous because the custom theme only waits for the SCSS files to be copied over.

This changes the watcher to listen for changes in SCSS files instead of CSS files.