chore: make provider factories docs-private
Most of the provider factories are already marked as `@docs-private`. This adds the `@docs-private` JSDoc tag to the missing ones.

This is necessary to avoid that the provider factories show up in the docs (once we have docs for the CDK) or in general when functions are also shown as part of a docs entry.