Please start providing context for nuanced config settings.
Every config I have seen lacks context.  When dealing with complex configs, we need to start providing the settings around the configuration options of interest in case other passer-bys are looking for where to include these specific and nuanced settings.  This is the only page that talks about scss being used in angular-cli.json