chore: fix invalid paths for codeowner folders
The GitHub code owners feature does not work properly right now because the different paths/globs to the folders (like the `input`)  are not matching at all.

This is because the paths are not absolute nor relative paths but still include a directory name. This does not work properly and can be confirmed by playing with the `.gitignore` (which has the same pattern matching)