chore(): fix package dependencies for screenshot-tool app
Recently some utilities have been moved from the Material package to the CDK package. At this point Material added the CDK as a peer dependency.

The screenshot-test application uses the most recent (released) Angular Material version (due to the caret in the semver expression) but doesn't have a dependency on the CDK yet.

This commit adds the CDK as a dependency and also switches Material and the CDK to the latest build artifcats on GitHub (as same as for the dashboard).