fix(expansion-panel): remove margin from top and bottom panels in accordion
Currently when the top-most or bottom-most expansion panels in an accordion are opened, they will add a top/bottom margin, causing them to shift away from the accordion edge. Since we try to avoid adding margins to the outside of components in general, these changes remove the top/bottom margins for the end panels.