fix(select): incorrect panel width when trigger is resized post initialization
Since we set and cache the trigger width in the `placeholder` setter, it means that we can potentially have an incorrect width if the trigger was resized after initialization. These changes switch to only updating the trigger width from the placeholder getter if there was a width already, which also has the advantage of not triggering a reflow on initialization, but deferring it until a select is opened.

Fixes #6655.