refactor(scroll-dispatcher): remove the need to pass in a callback to the scrolled method
Previously we required the consumer to pass in the subscription callback, when subscribing to `scrolled`, so we could manage the global resize handler. These changes refactor that functionality not to require the callback. This aligns the API with all of the other Observable-based APIs.

BREAKING CHANGE: Any uses of the `ScrollDispatcher.scrolled` method have to be refactored to subscribe to the returned Observable, instead of passing in the `callback`. Example

```ts
// Before
scrollDispatcher.scrolled(50, () => ...);

// After
scrollDispatcher.scrolled(50).subscribe(() => ...);
```