fix(button-toggle): border radius ignored if option is selected
As per Material Design specifications, the button-toggle group should always have a border radius.

This behavior has been implemented in the `md-button-toggle-group` component, but does not work properly if an option is selected or disabled. This happens because selected and disabled button-toggles receives a background color, which overflows the clipped button-toggle-group.

Fixes #6689